requirements.txt
ipykernel
langchain
langgraph
langchain-groq
langchain-community
langchain-core
python-dotenv
flask
ddgs
fpdf2
gunicorn